src/test/java/com/ideationstorm/com/ideationstorm/category/CategoryServiceTest.java
package com.ideationstorm.com.ideationstorm.category;

import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.ArgumentCaptor;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;

import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
import static org.mockito.Mockito.verify;

@ExtendWith(MockitoExtension.class)
class CategoryServiceTest {

    @Mock
    private CategoryRepository categoryRepository;
    private CategoryService underTest;


    @BeforeEach
    void setUp() {
        underTest = new CategoryService(categoryRepository);
    }

    @Test
    void updateCategory() {
        CategoryUpdateRequest updateRequest = CategoryUpdateRequest.builder().id(0).name("update").build();

        underTest.updateCategory(updateRequest);

        ArgumentCaptor<Category> categoryArgumentCaptor = ArgumentCaptor.forClass(Category.class);

        verify(categoryRepository).save(categoryArgumentCaptor.capture());

        Category capturedCategory = categoryArgumentCaptor.getValue();

        assertThat(capturedCategory.getId()).isEqualTo(updateRequest.getId());
        assertThat(capturedCategory.getName()).isEqualTo(updateRequest.getName());
    }
}